Update squawk_box_v1.ino
Added new SD loading functions.  The repeated code has been turned into its own function called fill_from_SD().  This function takes the filename to open as a parameter and returns whatever was in the file as a String.  This function is called multiple times in the loadContacts() function.  The loadContacts() function is still messy and needs some cleaning up, but it's more readable than before.  Strings returned from fill_from_SD() are converted into char arrays to be used in the sketch.