original DEV
SIM module does not reliably receive its own text data back from Twilio after Twilio send it back. It fails to read the data in the SMSrequest(). 
It does it the first time but then sometimes fails the second time and then just tries to reboot the SIM. 
NOT sure why...